Finalize test_core_gates
