+++ src/qibo/tests_new/test_core_gates_features.py
@@ -1,0 +1,31 @@
+"""Test special features of core gates."""
+import pytest
+import numpy as np
+import qibo
+from qibo import K, gates
+
+
+def test_rx_parameter_setter(backend):
+    """Check that the parameter setter of RX gate is working properly."""
+    def exact_state(theta):
+        phase = np.exp(1j * theta / 2.0)
+        gate = np.array([[phase.real, -1j * phase.imag],
+                        [-1j * phase.imag, phase.real]])
+        return gate.dot(np.ones(2)) / np.sqrt(2)
+
+    original_backend = qibo.get_backend()
+    qibo.set_backend(backend)
+    theta = 0.1234
+    gate = gates.RX(0, theta=theta)
+    initial_state = K.cast(np.ones(2) / np.sqrt(2))
+    final_state = gate(initial_state)
+    target_state = exact_state(theta)
+    np.testing.assert_allclose(final_state, target_state)
+
+    theta = 0.4321
+    gate.parameters = theta
+    initial_state = K.cast(np.ones(2) / np.sqrt(2))
+    final_state = gate(initial_state)
+    target_state = exact_state(theta)
+    np.testing.assert_allclose(final_state, target_state)
+    qibo.set_backend(original_backend)
